requirements.txt
django
django-guardian
djangorestframework